delete more awml pred files
