+++ autoware_mtr/dataclass/static_map.py
