+++ awml_pred/dataclass/__init__.py
@@ -1,6 +1,1 @@
-from .agent import *  # noqa
-from .polyline import *  # noqa
-from .prediction import *  # noqa
-from .scenario import *  # noqa
-from .static_map import *  # noqa
 from .utils import *  # noqa
