--- awml_pred/datasets/__init__.py
@@ -1,2 +1,0 @@
-from .base import *  # noqa
-from .builder import *  # noqa
