--- awml_pred/datasets/transforms/__init__.py
@@ -1,7 +1,0 @@
-from .agent import *  # noqa
-from .compose import *  # noqa
-from .filter import *  # noqa
-from .load import *  # noqa
-from .pad import *  # noqa
-from .polyline import *  # noqa
-from .utils import *  # noqa
