+++ projects/MTR/mtr/__init__.py
@@ -1,3 +1,2 @@
-from .datasets import *  # noqa
 from .models import *  # noqa
 from .ops import *   # noqa
