--- projects/MTR/mtr/datasets/__init__.py
@@ -1,2 +1,0 @@
-from .mtr_dataset import *  # noqa
-from .mtr_transform import *  # noqa
